Add style to nav and lists of books
