+++ src/App.scss
@@ -3,7 +3,8 @@
 @import './SCSS/books';
 
 body {
+  width: 100rem;
+  height: 89.625rem;
+  padding: 9.313rem 3.75rem 7.688rem 6.25rem;
   background-color: #f5f6fa;
-  margin: 0;
-  padding: 0;
 }
