+++ src/SCSS/_books.scss
@@ -1,18 +1,18 @@
 .books {
-  width: 83%;
-  margin: 0 auto;
-
   ul {
     list-style: none;
     margin: 0;
     padding: 0;
 
     li {
-      background-color: #fff;
-      margin-top: 10px;
+      width: 50rem;
       display: flex;
       justify-content: space-between;
-      padding: 10px;
+      padding: 1rem 5.188rem 1.625rem 1.688rem;
+      border-radius: 4px;
+      border: solid 1px #e8e8e8;
+      background-color: #fff;
+      margin: 0 auto;
     }
 
     .interraction {
