+++ src/SCSS/_bookscontainer.scss
@@ -1,7 +1,15 @@
 .container {
+  // width: 65rem;
+  // display: flex;
+  // flex-direction: column;
+  // // margin: 3rem auto;
+  // background-color: #fafafa;
   width: 65rem;
-  display: flex;
-  flex-direction: column;
-  margin: 3rem auto;
+  height: 68.563rem;
+  margin: 0 2.5rem 4.063rem 0;
+  padding: 0 0 49.625rem;
+  border-radius: 4px;
+  box-shadow: 0 0 60px 0 rgba(0, 0, 0, 0.05);
   background-color: #fafafa;
+  // border: 1px solid red;
 }
